Adds react-router to App and LandingPage Component
