+++ src/App.js
@@ -1,8 +1,17 @@
-import React from "react";
-import { LandingPage } from "./LandingPage/LandingPage";
+import React from 'react'
+import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom'
+import { LandingPage } from './LandingPage/LandingPage'
+import { ProductPage } from './ProductPage/ProductPage'
 
 function App() {
-  return <LandingPage />;
+  return (
+    <Router>
+      <Switch>
+        <Route path="/products" component={ProductPage} />
+        <Route path="/" exact component={LandingPage} />
+      </Switch>
+    </Router>
+  )
 }
 
-export default App;
+export default App
