+++ src/ProductPage/ProductPage.test.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import { render } from '@testing-library/react'
+import { ProductPage } from './ProductPage'
+
+it('should take a snapshot of LandingPage', () => {
+  const { asFragment } = render(<ProductPage />)
+  expect(asFragment(<ProductPage />)).toMatchSnapshot()
+})
